feat: 자동 로그인을 위한 임시 api를 추가한다
